Move declarations to correct files (related to sockets)
